+++ extensions/jsr/src/components/ItemDetails.tsx
@@ -1,143 +1,18 @@
-import { Color, Detail, Icon, List, open } from "@raycast/api";
-import { getProgressIcon } from "@raycast/utils";
+import { List } from "@raycast/api";
 
 import type { SearchResultDocument } from "@/types";
 
-import { compatIcons } from "@/lib/compat";
-import { jsrUrls } from "@/lib/jsrUrls";
-import { formatRelative, scoreColor } from "@/lib/ui-helpers";
+import { usePackage } from "@/hooks/jsrApi";
 
-import { useDependencies, useDependents, useDownloads, usePackage, usePackages, useVersions } from "@/hooks/jsrApi";
-
-import type { DepEntry } from "@/components/DepSection";
-import DepSection from "@/components/DepSection";
-import { useSearchContext } from "@/context/SearchContext";
-
-const MAX_DEPS_SHOWN = 10;
+import PackageMetadata from "@/components/PackageMetadata";
 
 const ItemDetails = ({ item }: { item: SearchResultDocument }) => {
-  const ctx = useSearchContext();
-  const icons = compatIcons(item);
-  const { data, isLoading } = usePackage(item);
-
-  const { data: scopePackages, isLoading: scopePackagesIsLoading } = usePackages(item.scope);
-  const { data: versionsData, isLoading: versionsIsLoading } = useVersions(isLoading ? null : (data ?? null));
-  const { data: downloadsData, isLoading: downloadsIsLoading } = useDownloads(isLoading ? null : (data ?? null));
-  const { data: dependentsData, isLoading: dependentsIsLoading } = useDependents(isLoading ? null : (data ?? null));
-  const { data: dependenciesData, isLoading: dependenciesIsLoading } = useDependencies(
-    isLoading ? null : (data ?? null),
-    data?.latestVersion ?? null,
-  );
-
-  const effectiveScore = data?.score ?? item.score ?? 0;
-  const effectiveColor = scoreColor(effectiveScore);
-
-  const latestPublishedAt = versionsData?.find((v) => v.version === data?.latestVersion)?.createdAt ?? null;
-  const latestPublishedText = versionsIsLoading ? "Loading…" : formatRelative(latestPublishedAt);
-
-  const totalDownloads = downloadsData?.total.reduce((sum, d) => sum + d.count, 0) ?? 0;
-  const totalDownloadsText = downloadsIsLoading ? "Loading…" : totalDownloads.toLocaleString();
-
-  const dependencyCount = data?.dependencyCount ?? dependenciesData?.length ?? 0;
-  const dependentCount = data?.dependentCount ?? dependentsData?.total ?? 0;
-
-  const dependencyEntries: DepEntry[] = (dependenciesData?.slice(0, MAX_DEPS_SHOWN) ?? []).map((dep) => {
-    const text = `${dep.kind}:${dep.name}${dep.path ? `/${dep.path}` : ""}`;
-    return {
-      key: text,
-      text,
-      icon: dep.kind === "jsr" ? "jsr.svg" : dep.kind === "npm" ? "npm.svg" : undefined,
-      onAction: () => {
-        if (dep.kind === "jsr") open(jsrUrls.site.package(dep.name));
-        else if (dep.kind === "npm") open(jsrUrls.site.npmPackage(dep.name));
-      },
-    };
-  });
-
-  const dependentEntries: DepEntry[] = (dependentsData?.items.slice(0, MAX_DEPS_SHOWN) ?? []).map((dep) => ({
-    key: dep.key,
-    text: `jsr:@${dep.scope}/${dep.package}`,
-    icon: "jsr.svg",
-    onAction: dep.package ? () => open(jsrUrls.site.scopePackage(dep.scope, dep.package as string)) : undefined,
-  }));
-
-  const hiddenDependencies = Math.max(0, dependencyCount - dependencyEntries.length);
-  const hiddenDependents = Math.max(0, dependentCount - dependentEntries.length);
-
+  const { isLoading } = usePackage(item);
   return (
     <List.Item.Detail
-      isLoading={
-        isLoading ||
-        dependentsIsLoading ||
-        dependenciesIsLoading ||
-        scopePackagesIsLoading ||
-        versionsIsLoading ||
-        downloadsIsLoading
-      }
+      isLoading={isLoading}
       markdown={[`## ${item.id}`, item.description].join("\n")}
-      metadata={
-        <Detail.Metadata>
-          {data ? (
-            <>
-              <Detail.Metadata.TagList title="Scope">
-                <Detail.Metadata.TagList.Item text={`@${item.scope}`} />
-                {typeof scopePackages?.total === "number" && scopePackages.total > 1 ? (
-                  <Detail.Metadata.TagList.Item
-                    text={`${scopePackages.items.length}`}
-                    icon={Icon.Box}
-                    onAction={() => ctx?.openScope(item.scope)}
-                  />
-                ) : null}
-                {data.isArchived ? (
-                  <Detail.Metadata.TagList.Item text="Archived" color={Color.Red} icon={Icon.XMarkCircle} />
-                ) : null}
-              </Detail.Metadata.TagList>
-              <Detail.Metadata.Label title="Last Published" text={latestPublishedText} icon={Icon.Clock} />
-              <Detail.Metadata.Label title="Created" text={formatRelative(data.createdAt)} icon={Icon.Calendar} />
-              <Detail.Metadata.Label
-                title="Version"
-                text={`${data.latestVersion ?? "unknown"}${data.versionCount ? ` (${data.versionCount} total)` : ""}`}
-                icon={Icon.ComputerChip}
-              />
-              <Detail.Metadata.Label title="Downloads (90d)" text={totalDownloadsText} icon={Icon.Download} />
-              <Detail.Metadata.Separator />
-            </>
-          ) : null}
-          <Detail.Metadata.Label
-            title="Score"
-            text={`${effectiveScore}%`}
-            icon={{ source: getProgressIcon(effectiveScore / 100, effectiveColor, { backgroundOpacity: 0 }) }}
-          />
-          <Detail.Metadata.TagList title="Compatibility">
-            {icons.map((ico) => (
-              <Detail.Metadata.TagList.Item key={ico.text} text={ico.text} icon={ico.icon} />
-            ))}
-          </Detail.Metadata.TagList>
-          <DepSection
-            title="Dependencies"
-            totalCount={dependencyCount}
-            entries={dependencyEntries}
-            hiddenCount={hiddenDependencies}
-            moreUrl={jsrUrls.site.packageDependencies(item.scope, item.name)}
-          />
-          <DepSection
-            title="Dependents"
-            totalCount={dependentCount}
-            entries={dependentEntries}
-            hiddenCount={hiddenDependents}
-            moreUrl={jsrUrls.site.packageDependents(item.scope, item.name)}
-          />
-          <Detail.Metadata.Separator />
-          <Detail.Metadata.Link title="JSR" text="View on jsr.io" target={jsrUrls.site.package(item.id)} />
-          {data?.githubRepository?.owner && data?.githubRepository?.name ? (
-            <Detail.Metadata.Link
-              title="GitHub"
-              text="View on GitHub"
-              target={jsrUrls.site.github(data.githubRepository.owner, data.githubRepository.name)}
-            />
-          ) : null}
-        </Detail.Metadata>
-      }
+      metadata={<PackageMetadata item={item} />}
     />
   );
 };
